Add .gitignore
- Add api file in routes folder
