+++ main.go
@@ -1,36 +1,8 @@
 package main
 
-import (
-	"net/http"
-
-	"example.com/m/controllers"
-	"example.com/m/models"
-	"github.com/gin-gonic/gin"
-)
-
-func setupRouter() *gin.Engine {
-	r := gin.Default()
-
-	models.ConnectionDatabase()
-
-	r.GET("/", func(c *gin.Context) {
-		c.JSON(http.StatusOK, gin.H{"data": "Hello, world!"})
-	})
-
-	r.GET("/mastheads", controllers.GetMastheads)
-
-	r.POST("/mastheads", controllers.CreateMasthead)
-
-	r.GET("/mastheads/:id", controllers.GetMasthead)
-
-	r.PATCH("/mastheads/:id", controllers.UpdateMasthead)
-
-	r.DELETE("/mastheads/:id", controllers.DeleteMasthead)
-
-	return r
-}
+import "example.com/m/routes"
 
 func main() {
-	r := setupRouter()
+	r := routes.Router()
 	r.Run(":3000")
 }
